Created CSS and Updated HTML
Styled body, h1, wrapped from header to  main section into a <div class= container>, text-aligned h1, worked on the header, styled the form.
Added a class tp the input and styled it in CSS., added class to the submit button and styled it,  added shadow on  the form and styled it some more. Added class to the div holding the poem and styled it . Styled the footer. Colored all the links with the same color to make it more consistent.